+++ src/main/java/com/part4/team05/sb01otbooteam05/domain/feedComment/service/FeedCommentService.java
@@ -1,5 +1,16 @@
 package com.part4.team05.sb01otbooteam05.domain.feedComment.service;
 
-public class FeedCommentService {
+import java.util.UUID;
+
+import com.part4.team05.sb01otbooteam05.domain.feedComment.dto.CommentDto;
+import com.part4.team05.sb01otbooteam05.domain.feedComment.dto.CommentDtoCursorResponse;
+import com.part4.team05.sb01otbooteam05.domain.feedComment.dto.request.CommentCreateRequest;
+import com.part4.team05.sb01otbooteam05.domain.feedComment.dto.request.FindCommentsRequest;
+
+public interface FeedCommentService {
+
+	CommentDtoCursorResponse findComments(UUID userId, FindCommentsRequest request);
+
+	CommentDto createFeedComment(UUID userId, UUID feedId, CommentCreateRequest request);
 
 }
